Restructure form pages
